MB-6237 Move MilMoveDomain and Env enums into hosts.py
