+++ locustfiles/prime.py
@@ -2,8 +2,8 @@
 """ Locust test for the Prime & Support APIs """
 from locust import HttpUser, between
 
-from utils.constants import MilMoveDomain, LOCAL_TLS_CERT_KWARGS
-from utils.hosts import MilMoveHostMixin
+from utils.constants import LOCAL_TLS_CERT_KWARGS
+from utils.hosts import MilMoveHostMixin, MilMoveDomain
 from utils.parsers import PrimeAPIParser, SupportAPIParser
 from tasks import PrimeTasks, SupportTasks
